updated @actions/core dep to 1.10.0
Updated the @actions/core dep to 1.10.0 in order to resolve the warnings
tied to the deprecated GitHub action `set-output` command. This new
version using environment files for std output and is more secure. The
previous version is deprecated and will be removed in March 2023.

More details can be found here: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>